+++ SOUP.user.js
@@ -1,9 +1,9 @@
 // ==UserScript==
 // @name        Stack Overflow Unofficial Patch
 // @namespace   https://github.com/vyznev/
-// @description Miscellaneous client-side fixes for bugs on Stack Exchange sites (development)
+// @description Miscellaneous client-side fixes for bugs on Stack Exchange sites
 // @author      Ilmari Karonen
-// @version     1.37.5
+// @version     1.38.0
 // @copyright   2014-2015, Ilmari Karonen (http://stackapps.com/users/10283/ilmari-karonen)
 // @license     ISC; http://opensource.org/licenses/ISC
 // @match       *://*.stackexchange.com/*
@@ -14,9 +14,9 @@
 // @match       *://*.mathoverflow.net/*
 // @match       *://*.askubuntu.com/*
 // @homepageURL http://stackapps.com/questions/4486/stack-overflow-unofficial-patch
-// @updateURL   https://github.com/vyznev/soup/raw/devel/SOUP.meta.js
-// @downloadURL https://github.com/vyznev/soup/raw/devel/SOUP.user.js
-// @icon        https://github.com/vyznev/soup/raw/devel/icon/SOUP_icon_128.png
+// @updateURL   https://github.com/vyznev/soup/raw/master/SOUP.meta.js
+// @downloadURL https://github.com/vyznev/soup/raw/master/SOUP.user.js
+// @icon        https://github.com/vyznev/soup/raw/master/icon/SOUP_icon_128.png
 // @grant       none
 // @run-at      document-start
 // ==/UserScript==